Implement reporter
