+++ newsroom/src/main/java/com/theguardian/newsroom/Newsroom.kt
@@ -6,59 +6,33 @@ import android.content.Context
 import android.os.Build
 import android.support.v4.app.NotificationCompat
 import android.support.v4.app.NotificationManagerCompat
-import android.util.Log
-import io.reactivex.Observable
-import io.reactivex.disposables.Disposable
-import io.reactivex.schedulers.Schedulers
+import com.theguardian.newsroom.model.Event
 import java.util.concurrent.atomic.AtomicInteger
 
-object Newsroom {
+class Newsroom(private val context: Context) {
 
-    private const val TAG = "Newsroom"
-    private const val NOTIFICATION_CHANNEL_ID = "newsroom"
+    private val notificationManager = NotificationManagerCompat.from(context)
+    private val notificationId = AtomicInteger(1821)
 
-    private fun logcat(options: String): Observable<String> {
-        val observable: Observable<String> = Observable.create { emitter ->
-            val process = Runtime.getRuntime().exec("logcat $options")
-            val bufferedReader = process.inputStream.reader().buffered()
-
-            emitter.setCancellable {
-                bufferedReader.close()
-                process.destroy()
-            }
-
-            // Note: forEachLine closes the reader automatically
-            bufferedReader.forEachLine(emitter::onNext)
-
-            process.destroy()
-            emitter.onComplete()
-        }
-        return observable.subscribeOn(Schedulers.io())
+    fun reportEvent(event: Event){
+        notification(event)
     }
 
-    private fun gaLogToMap(logLine: String): Map<String, String> {
-        val result = mutableMapOf<String, String>()
-        val regex = Regex(", [a-z0-9_]+=")
-        var start = 0
-        var match = regex.find(logLine, start)
-        while (match != null) {
-            val split = logLine.subSequence(start, match.range.start).split("=", limit = 2)
-            val key = split[0]
-            val value = split[1]
-            result[key] = value
-            start = match.range.start + 2
-            match = regex.find(logLine, start)
-        }
-        val split = logLine.subSequence(start, logLine.length).split("=", limit = 2)
-        val key = split[0]
-        val value = split[1]
-        result[key] = value
-        return result
+    fun addReporter(reporter: Reporter): Newsroom {
+        reporter.setNewsroom(this)
+        reporter.onStart()
+        return this
     }
 
-    private fun gaHitDeliveries(): Observable<String> = logcat("-s GAv4")
-            .filter { it.contains("Hit delivery requested") }
-            .map { it.split("Hit delivery requested:").last() }
+    private fun notification(event: Event){
+        val notification = newNotification(context)
+                .setContentTitle(event.title)
+                .setContentText(event.message)
+                .setStyle(NotificationCompat.BigTextStyle().bigText(event.message).setSummaryText("Expand for details"))
+                .setSmallIcon(android.R.drawable.stat_notify_error)
+                .build()
+        notificationManager.notify(notificationId.getAndIncrement(), notification)
+    }
 
     private fun newNotification(context: Context): NotificationCompat.Builder {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
@@ -68,22 +42,7 @@ object Newsroom {
         return NotificationCompat.Builder(context, NOTIFICATION_CHANNEL_ID)
     }
 
-    fun notifyWhenGaHitsAreSent(context: Context): Disposable {
-        val notificationManager = NotificationManagerCompat.from(context)
-        val notificationId = AtomicInteger(1821)
-        return gaHitDeliveries()
-                .map { gaLogToMap(it) }
-                .subscribe({ map ->
-                    if (map["t"] == "event") {
-                        val message = "Category: ${map["ec"]}\nAction: ${map["ea"]}\nLabel: ${map["el"]}"
-                        val notification = newNotification(context)
-                                .setContentTitle("GA Event Tracked")
-                                .setContentText(message)
-                                .setStyle(NotificationCompat.BigTextStyle().bigText(message).setSummaryText("Expand for details"))
-                                .setSmallIcon(android.R.drawable.stat_notify_error)
-                                .build()
-                        notificationManager.notify(notificationId.getAndIncrement(), notification)
-                    }
-                }, { err -> Log.w(TAG, err) })
+    companion object {
+        private const val NOTIFICATION_CHANNEL_ID = "newsroom"
     }
 }
